chore: Move file related to definition of application errors
