+++ src/libs/error/applicationErrors.ts
